+++ packages/motion-core/src/lib/components/god-rays/GodRaysScene.svelte
@@ -109,7 +109,7 @@
 	}: Props = $props();
 
 	let canvas = $state<HTMLCanvasElement>();
-	let uniforms = $state<{
+	let uniforms = $state.raw<{
 		uTime: { value: number };
 		uResolution: { value: Vec2 };
 		uColor: { value: Vec3 };