+++ packages/motion-core/src/lib/components/halo/HaloScene.svelte
@@ -67,7 +67,7 @@
 	}: Props = $props();
 
 	let canvas = $state<HTMLCanvasElement>();
-	let uniforms = $state<{
+	let uniforms = $state.raw<{
 		uTime: { value: number };
 		uResolution: { value: Vec2 };
 		uBackgroundColor: { value: Vec3 };